requirements.txt
# FastAPI and server dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Computer Vision and ML
opencv-python==4.9.0.80
opencv-contrib-python==4.9.0.80
ultralytics==8.1.0
torch==2.1.2
torchvision==0.16.2
pillow==10.2.0
numpy==1.26.3

# OCR
pytesseract==0.3.10

# Image processing
scikit-image==0.22.0
scipy==1.11.4

# Utilities
python-dotenv==1.0.0
aiofiles==23.2.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0

# Logging and monitoring
loguru==0.7.2
